updated anotheria- and foreign dependencies and did a lot of little refactorings to improve performance and code quality

+++ distributeme-test/src/main/java/org/distributeme/test/interception/interceptor/InterceptSumInterceptor.java
@@ -25,8 +25,8 @@ public class InterceptSumInterceptor extends AbstractClientSideRequestIntercepto
 		List parameters = context.getParameters();
 		Integer a = (Integer) parameters.get(0);
 		Integer b = (Integer) parameters.get(1);
-		parameters.set(0, Integer.valueOf(a * 2));
-		parameters.set(1, Integer.valueOf(b * 2));
+		parameters.set(0, a * 2);
+		parameters.set(1, b * 2);
 		//this call is not needed, but it makes the code readable.
 		context.setParameters(parameters);
 		
@@ -46,7 +46,7 @@ public class InterceptSumInterceptor extends AbstractClientSideRequestIntercepto
 		}
 		
 		Integer ret = (Integer)iContext.getReturnValue();
-		return InterceptorResponse.returnNow(ret.intValue()*3);		
+		return InterceptorResponse.returnNow(ret *3);
 		
 	}
